añadir cobro y añadir excusa ; componentes

+++ app/vistas/auth/deudor/solicitudes-deudor.php
@@ -1,6 +1,9 @@
 <main>
     <h2 class="vista-titulo">Historial de Prestamos</h2>
-    <table class="table table-striped text-center">
+    <form action="" method="get">
+        <?php FilterOptionsComponent() ?>
+    </form>
+    <table class="table mt-0 table-striped text-center">
         <thead>
             <tr>
                 <th scope="col">Pagodiario</th>
@@ -13,27 +16,31 @@
         </thead>
         <tbody>
             <?php
-            foreach ($prestamos as $row) { ?>
-                <tr>
-                    <th scope="row"><?php echo $row['pagodiario'] ?></th>
-                    <td scope="row">$<?php echo $row['cantidad'] ?></td>
-                    <td>$<?php echo $row['deuda'] ?></td>
-                    <td><?php echo date("d-m-Y", strtotime($row['fecha_solicitud'])) ?></td>
-                    <td <?php if ($row["estado"] != "Aceptada") { ?>class="row" <?php } ?>>
-                        <?php if ($row["estado"] != "Aceptada") { ?>
-                            <a href="/deudor/mis-solicitudes/delete?soli=<?php echo $row["id_prestamo"] ?>" class="btn btn-danger">
-                                <i class="fa fa-trash"></i>
-                            </a>
-                        <?php } ?>
-                    </td>
-                    <td class="p-2">
-                        <div class="row px-3">
-                            <span class="estado <?php echo strtolower($row['estado']) ?> py-2 rounded"><?php echo ($row['estado']) ?>
-                            </span>
-                        </div>
-                    </td>
-                </tr>
-            <?php } ?>
+            if (count($prestamos)) {
+                foreach ($prestamos as $row) { ?>
+                    <tr>
+                        <th scope="row"><?php echo $row['pagodiario'] ?></th>
+                        <td scope="row">$<?php echo $row['cantidad'] ?></td>
+                        <td>$<?php echo $row['deuda'] ?></td>
+                        <td><?php echo date("d-m-Y", strtotime($row['fecha_solicitud'])) ?></td>
+                        <td <?php if ($row["estado"] != "Aceptada") { ?>class="row" <?php } ?>>
+                            <?php if ($row["estado"] != "Aceptada") { ?>
+                                <a href="/deudor/mis-solicitudes/delete?soli=<?php echo $row["id_prestamo"] ?>" class="btn btn-danger">
+                                    <i class="fa fa-trash"></i>
+                                </a>
+                            <?php } ?>
+                        </td>
+                        <td class="p-2">
+                            <div class="row px-3">
+                                <span class="estado <?php echo strtolower($row['estado']) ?> py-2 rounded"><?php echo ($row['estado']) ?>
+                                </span>
+                            </div>
+                        </td>
+                    </tr>
+            <?php }
+            } else {
+                NoRegistrosComponent();
+            } ?>
         </tbody>
     </table>
 </main>
